+++ src/components/layout/dashboard/LeftNavbarDashboard.tsx
@@ -1,0 +1,94 @@
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipTrigger,
+  TooltipProvider,
+} from "@/components/ui/tooltip";
+import Link from "next/link";
+import {
+  File,
+  Home,
+  Settings,
+  Users2,
+  ChartColumnDecreasing,
+} from "lucide-react";
+
+export default function LeftNavbarDashboard() {
+  return (
+    <div className="min-h-screen">
+      <aside className="fixed z-10 hidden w-14 flex-col border-r bg-background sm:flex">
+        <nav className="flex flex-col items-center">
+          <TooltipProvider>
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Link
+                  href="#"
+                  className="flex h-9 w-9 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:text-foreground md:h-8 md:w-8"
+                >
+                  <Home className="h-5 w-5" />
+                  <span className="sr-only">Tablero</span>
+                </Link>
+              </TooltipTrigger>
+              <TooltipContent side="right">Tablero</TooltipContent>
+            </Tooltip>
+
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Link
+                  href="#"
+                  className="flex h-9 w-9 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:text-foreground md:h-8 md:w-8"
+                >
+                  <File className="h-5 w-5" />
+                  <span className="sr-only">Archivos</span>
+                </Link>
+              </TooltipTrigger>
+              <TooltipContent side="right">Archivos</TooltipContent>
+            </Tooltip>
+
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Link
+                  href="#"
+                  className="flex h-9 w-9 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:text-foreground md:h-8 md:w-8"
+                >
+                  <Users2 className="h-5 w-5" />
+                  <span className="sr-only">Usuarios</span>
+                </Link>
+              </TooltipTrigger>
+              <TooltipContent side="right">Usuarios</TooltipContent>
+            </Tooltip>
+
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Link
+                  href="#"
+                  className="flex h-9 w-9 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:text-foreground md:h-8 md:w-8"
+                >
+                  <ChartColumnDecreasing className="h-5 w-5" />
+                  <span className="sr-only">Estadisticas</span>
+                </Link>
+              </TooltipTrigger>
+              <TooltipContent side="right">Estadisticas</TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+        </nav>
+        <nav className="mt-auto flex flex-col items-center gap-4 px-2 py-4">
+          <TooltipProvider>
+            <Tooltip>
+              <TooltipTrigger asChild>
+                <Link
+                  href="#"
+                  className="flex h-9 w-9 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:text-foreground md:h-8 md:w-8"
+                >
+                  <Settings className="h-5 w-5" />
+                  <span className="sr-only">Configuración</span>
+                </Link>
+              </TooltipTrigger>
+              <TooltipContent side="right">Configuración</TooltipContent>
+            </Tooltip>
+          </TooltipProvider>
+        </nav>
+      </aside>
+    </div>
+  );
+}
