+++ src/pages/file/index.tsx
@@ -1,0 +1,10 @@
+import { Dashboard } from "@/components/layout/dashboard/Dashboard";
+import React from "react";
+
+export default function FilePage() {
+  return (
+    <section className="flex items-center justify-center my-4 p-2 gap-4 sm:p-4">
+      <Dashboard />
+    </section>
+  );
+}
